refactor(content): WKC layout — Phase 5 (active-entities, bloom filter, pointers)
Phase 5 of the WKC migration. Reshapes the active-entities subsystem and
the surrounding pointer/bloom-filter components, finishing the split
that was deferred from Phase 2.

New / moved components:

- adapters/active-entities-repository/ — new. Owns the SQL that touches
  the active_pointers projection (updateActiveDeployments,
  removeActiveDeployments). Extracted from the pointers-repository
  created in Phase 2 — those two functions write to active_pointers,
  not to a pointer-history table, so they belong with the active
  entities, not with pointer history reads.
- adapters/deployed-entities-bloom-filter/ — moved from
  ports/deployedEntitiesBloomFilter.ts. The probabilistic membership
  filter is rules-free state, so it stays an adapter (per merge review).
- logic/active-entities/ — moved from ports/activeEntities.ts (393 LoC).
  This was the largest single module split: it now consumes the
  active-entities-repository (writes) and pointers-repository (URN-prefix
  reads) instead of importing SQL functions directly. Same public surface
  (withPointers, withIds, withPrefix, update, clear, clearPointers,
  getCachedEntity, reset). Types and constants moved to types.ts.
- logic/pointer-manager/ — moved from service/pointers/PointerManager.ts.
  Kept as a class (the test helper uses ts-mockito's mock(PointerManager)
  which needs the class shape). DELTA_POINTER_RESULT and DeploymentResult
  moved to types.ts.

Folded into existing components:

- service/pointers/pointers.ts (getPointerChanges) -> logic/deployments/.
  It's a paginated read over historical deployments with denylist
  filtering — the same shape as logic/deployments/getDeployments. Moves
  the function plus its DeploymentPointerChanges and PointerChangesFilters
  types into the deployments component.

Repository surface change:

- pointers-repository now exposes only URN-prefix lookups
  (getItemEntitiesIdsThatMatchCollectionUrnPrefix,
  getThirdPartyCollectionItemsEntityIdsThatMatchUrnPrefix). The
  active_pointers writes that lived here transitionally moved to the new
  active-entities-repository.

Wiring:

- AppComponents gets activeEntitiesRepository.
- src/service/pointers/ deleted entirely.
- All importers updated; module-level function exports preserved so
  jest.spyOn(...) patterns keep working.

tsc clean, unit suite green (213 tests, no diff).
