+++ content/src/adapters/active-entities-repository/index.ts
@@ -1,0 +1,2 @@
+export { createActiveEntitiesRepository, updateActiveDeployments, removeActiveDeployments } from './component'
+export type { IActiveEntitiesRepository } from './types'
