+++ content/src/adapters/deployed-entities-bloom-filter/index.ts
@@ -1,0 +1,2 @@
+export { createDeployedEntitiesBloomFilter } from './component'
+export type { DeployedEntitiesBloomFilter } from './types'
