+++ content/src/controller/handlers/pointer-changes-handler.ts
@@ -4,8 +4,8 @@ import { asEnumValue, fromCamelCaseToSnakeCase } from '../utils'
 import { qsGetArray, qsGetBoolean, qsGetNumber, qsParser, toQueryParams } from '../../logic/query-params'
 import { HandlerContextWithPath, InvalidRequestError, parseEntityType } from '../../types'
 import { SortingField, SortingOrder } from '../../deployment-types'
-import { PointerChangesFilters } from '../../service/pointers/types'
-import { getPointerChanges } from '../../service/pointers/pointers'
+import { PointerChangesFilters } from '../../logic/deployments'
+import { getPointerChanges } from '../../logic/deployments'
 
 // Method: GET
 // Query String: ?from={timestamp}&to={timestamp}&offset={number}&limit={number}&entityType={entityType}&includeAuthChain={boolean}
