+++ content/src/logic/active-entities/index.ts
@@ -1,0 +1,3 @@
+export { createActiveEntitiesComponent } from './component'
+export { BASE_AVATARS_COLLECTION_ID, BASE_EMOTES_COLLECTION_ID, isEntityPresent, isPointingToEntity } from './types'
+export type { ActiveEntities, NotActiveEntity } from './types'
