+++ content/src/logic/entities.ts
@@ -1,6 +1,6 @@
 import { Entity } from '@dcl/schemas'
 import { IDatabaseComponent } from 'src/ports/postgres'
-import { ActiveEntities } from '../ports/activeEntities'
+import { ActiveEntities } from '../logic/active-entities'
 
 export async function findEntityByPointer(
   database: IDatabaseComponent,
