+++ content/src/logic/pointer-manager/index.ts
@@ -1,0 +1,3 @@
+export { PointerManager } from './component'
+export { DELTA_POINTER_RESULT } from './types'
+export type { DeploymentResult } from './types'
