--- content/src/service/pointers/pointers.ts
@@ -1,45 +1,0 @@
-import { PointerChangesSyncDeployment } from '@dcl/schemas'
-import { DatabaseClient } from 'src/ports/postgres'
-import { PointerChangesOptions } from '../../deployment-types'
-import { getHistoricalDeployments, HistoricalDeployment } from '../../adapters/deployments-repository'
-import { AppComponents } from '../../types'
-import { DeploymentPointerChanges } from './types'
-
-const MAX_HISTORY_LIMIT = 500
-
-export async function getPointerChanges(
-  components: Pick<AppComponents, 'denylist' | 'metrics'>,
-  database: DatabaseClient,
-  options?: PointerChangesOptions
-): Promise<DeploymentPointerChanges> {
-  const curatedOffset = options?.offset && options?.offset >= 0 ? options?.offset : 0
-  const curatedLimit =
-    options?.limit && options?.limit > 0 && options?.limit <= MAX_HISTORY_LIMIT ? options?.limit : MAX_HISTORY_LIMIT
-  let deploymentsWithExtra: HistoricalDeployment[] = await getHistoricalDeployments(
-    database,
-    curatedOffset,
-    curatedLimit + 1,
-    options?.filters,
-    options?.sortBy,
-    options?.lastId
-  )
-
-  // Note: moreData is checked before denylist filtering so pagination signals remain correct.
-  // This means returned pages may have fewer than curatedLimit items when denylisted entities
-  // are removed. Fixing that would require fetching additional rows, left as a future improvement.
-  const moreData: boolean = deploymentsWithExtra.length > curatedLimit
-  deploymentsWithExtra = deploymentsWithExtra.filter((result) => !components.denylist.isDenylisted(result.entityId))
-  const deployments: PointerChangesSyncDeployment[] = deploymentsWithExtra.slice(0, curatedLimit)
-
-  return {
-    pointerChanges: deployments,
-    filters: {
-      ...options?.filters
-    },
-    pagination: {
-      offset: curatedOffset,
-      limit: curatedLimit,
-      moreData
-    }
-  }
-}
