--- content/src/service/pointers/types.ts
@@ -1,16 +1,0 @@
-import { PointerChangesSyncDeployment } from '@dcl/schemas'
-import { DeploymentFilters } from '../../deployment-types'
-
-export type PointerChangesFilters = Pick<DeploymentFilters, 'from' | 'to' | 'entityTypes'>
-
-export type DeploymentPointerChanges = {
-  pointerChanges: PointerChangesSyncDeployment[]
-  filters: Omit<PointerChangesFilters, 'entityType'>
-  pagination: {
-    offset: number
-    limit: number
-    moreData: boolean
-    lastId?: string
-    next?: string
-  }
-}
