+++ content/test/integration/controller/deployment-pagination.spec.ts
@@ -3,7 +3,7 @@ import { createFetchComponent } from '@well-known-components/fetch-component'
 import { DeploymentField } from '../../../src/types'
 import { DeploymentOptions, SortingField, SortingOrder } from '../../../src/deployment-types'
 import { toQueryParams } from '../../../src/logic/query-params'
-import { PointerChangesFilters } from '../../../src/service/pointers/types'
+import { PointerChangesFilters } from '../../../src/logic/deployments'
 import { makeNoopValidator } from '../../helpers/service/validations/NoOpValidator'
 import { buildDeployData, EntityCombo } from '../E2ETestUtils'
 import { resetServer, createDefaultServer } from '../simpleTestEnvironment'
