+++ content/test/unit/controllers/filter-by-urn-handler.spec.ts
@@ -1,7 +1,7 @@
 import { createMockedActiveEntitiesComponent } from '../../mocks/active-entities-component-mock'
 import { getEntitiesByCollectionPointerPrefixHandler } from '../../../src/controller/handlers/filter-by-urn-handler'
 import { HandlerContextWithPath, InvalidRequestError } from '../../../src/types'
-import { ActiveEntities } from '../../../src/ports/activeEntities'
+import { ActiveEntities } from '../../../src/logic/active-entities'
 import { createMockedEntity } from '../../mocks/entity-mock'
 import { Entity } from '@dcl/schemas'
