+++ content/test/unit/ports/activeEntities.spec.ts
@@ -10,9 +10,9 @@ import { Deployment } from '../../../src/deployment-types'
 import { DEFAULT_ENTITIES_CACHE_SIZE, Environment, EnvironmentConfig } from '../../../src/Environment'
 import * as deployments from '../../../src/logic/deployments'
 import { metricsDeclaration } from '../../../src/metrics'
-import { ActiveEntities, createActiveEntitiesComponent } from '../../../src/ports/activeEntities'
+import { ActiveEntities, createActiveEntitiesComponent } from '../../../src/logic/active-entities'
 import { Denylist } from '../../../src/adapters/denylist'
-import { createDeployedEntitiesBloomFilter } from '../../../src/ports/deployedEntitiesBloomFilter'
+import { createDeployedEntitiesBloomFilter } from '../../../src/adapters/deployed-entities-bloom-filter'
 import { createNoOpDeployRateLimiter } from '../../mocks/deploy-rate-limiter-mock'
 import { createFailedDeployments } from '../../../src/ports/failedDeployments'
 import { createAuthenticator } from '../../../src/logic/authenticator'
