+++ content/test/unit/ports/deployer.spec.ts
@@ -19,20 +19,20 @@ import {
 } from '../../../src/deployment-types'
 import * as deploymentQueries from '../../../src/adapters/deployments-repository'
 import * as failedDeploymentQueries from '../../../src/adapters/failed-deployments-repository'
-import * as pointers from '../../../src/adapters/pointers-repository'
+import * as pointers from '../../../src/adapters/active-entities-repository'
 import * as deploymentLogic from '../../../src/logic/deployments'
 import * as deployments from '../../../src/logic/deployments'
 import { metricsDeclaration } from '../../../src/metrics'
-import { createActiveEntitiesComponent } from '../../../src/ports/activeEntities'
+import { createActiveEntitiesComponent } from '../../../src/logic/active-entities'
 import { Denylist } from '../../../src/adapters/denylist'
 import { createNoOpDeployRateLimiter } from '../../mocks/deploy-rate-limiter-mock'
-import { createDeployedEntitiesBloomFilter } from '../../../src/ports/deployedEntitiesBloomFilter'
+import { createDeployedEntitiesBloomFilter } from '../../../src/adapters/deployed-entities-bloom-filter'
 import { createDeployer } from '../../../src/ports/deployer'
 import { createFailedDeployments } from '../../../src/ports/failedDeployments'
 import { createTestDatabaseComponent } from '../../../src/ports/postgres'
 import { createSequentialTaskExecutor } from '../../../src/adapters/sequential-task-executor'
 import { createAuthenticator } from '../../../src/logic/authenticator'
-import { DELTA_POINTER_RESULT } from '../../../src/service/pointers/PointerManager'
+import { DELTA_POINTER_RESULT } from '../../../src/logic/pointer-manager'
 import { EntityVersion } from '../../../src/types'
 import { buildEntityAndFile } from '../../helpers/entity-tests-helper'
 import { NoOpServerValidator, NoOpValidator } from '../../helpers/service/validations/NoOpValidator'
